+++ projects/ui-framework/src/lib/indicators/progress/progress-bar/progress-bar.component.html
