+++ projects/ui-framework/src/lib/indicators/progress/progress-bar/progress-bar.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProgressBarComponent } from './progress-bar.component';
-import { UtilsService } from '../../services/utils/utils.service';
-import { DOMhelpers } from '../../services/html/dom-helpers.service';
+import { UtilsService } from '../../../services/utils/utils.service';
+import { DOMhelpers } from '../../../services/html/dom-helpers.service';
 
 @NgModule({
   declarations: [ProgressBarComponent],
   imports: [CommonModule],
   exports: [ProgressBarComponent],
-  providers: [UtilsService, DOMhelpers]
+  providers: [UtilsService, DOMhelpers],
 })
 export class ProgressBarModule {}
