+++ projects/ui-framework/src/lib/indicators/progress/progress-donut/progress-donut.component.html
@@ -1,0 +1,10 @@
+<svg class="bpd-donut"
+     [attr.viewbox]="'0 0 ' + diameter + ' ' + diameter">
+  <circle class="bpd-donut-circle"
+          [attr.cx]="diameter/2"
+          [attr.cy]="diameter/2"
+          [attr.r]="diameter/2"
+          [attr.stroke-dasharray]="circumference"
+          [attr.stroke-dashoffset]="circumference"
+          [attr.stroke-width]="stroke"></circle>
+</svg>
