+++ projects/ui-framework/src/lib/indicators/progress/progress-donut/progress-donut.const.ts
@@ -1,0 +1,7 @@
+import { ProgressDonutSize } from './progress-donut.enum';
+
+export const ProgressDonutDiameter = {
+  [ProgressDonutSize.small]: 32,
+  [ProgressDonutSize.medium]: 46,
+  [ProgressDonutSize.large]: 110,
+};
