+++ projects/ui-framework/src/lib/indicators/progress/progress-donut/progress-donut.enum.ts
@@ -1,8 +1,3 @@
-export enum ProgressDonutType {
-  primary = 'primary',
-  secondary = 'secondary',
-}
-
 export enum ProgressDonutSize {
   small = 'small',
   medium = 'medium',
