+++ projects/ui-framework/src/lib/indicators/progress/progress-donut/progress-donut.interface.ts
@@ -1,11 +1,8 @@
-import { Icons } from '../../icons/icons.enum';
-
 export interface ProgressDonutData {
   value: number;
   color?: string;
   headerTextPrimary?: string | boolean;
   headerTextSecondary?: string | boolean;
-  iconHeaderRight?: Icons;
 }
 
 export interface ProgressDonutConfig {
