+++ projects/ui-framework/src/lib/indicators/progress/progress-donut/progress-donut.stories.ts
